pyproject.toml
[tool.poetry]
name = "rag-sample"
version = "0.1.0"
description = ""
authors = ["piyota0901 <piyo.tatsu.0901@gmail.com>"]
readme = "README.md"
packages = [{ include = "mypackage" }]

[tool.poetry.dependencies]
python = "3.11.5"
qdrant-client = "^1.7.0"
sentence-transformers = "^2.2.2"
injector = "^0.21.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
python-dotenv = "^1.0.0"
openai = "^1.8.0"


[tool.poetry.group.dev.dependencies]
llama-cpp-python = "^0.2.30"


[tool.poetry.group.tools.dependencies]
pandas = "^2.1.4"
llama-cpp-python = {extras = ["server"], version = "^0.2.31"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
